Refactoring: slimmed down fat module
