+++ leitus/luks.py
@@ -20,7 +20,8 @@
 #
 #
 
-from leitus.deep import LuksDevice, DiskByUUID
+from leitus.deep import LuksDevice
+from leitus.device import DiskByUUID
 
 INFO = """
